refactor: remove IMPL file
[#OCD-3831]
